Add Dockerfile (#55)

+++ tests/Ethereum/RLPTests.cpp
@@ -25,8 +25,8 @@ TEST(RLP, Integers) {
     EXPECT_EQ(hex(RLP::encode(256)), "820100");
     EXPECT_EQ(hex(RLP::encode(1024)), "820400");
     EXPECT_EQ(hex(RLP::encode(0xffffff)), "83ffffff");
-    EXPECT_EQ(hex(RLP::encode(0xffffffffULL)), "84ffffffff");
-    EXPECT_EQ(hex(RLP::encode(0xffffffffffffffULL)), "87ffffffffffffff");
+    EXPECT_EQ(hex(RLP::encode(static_cast<uint64_t>(0xffffffffULL))), "84ffffffff");
+    EXPECT_EQ(hex(RLP::encode(static_cast<uint64_t>(0xffffffffffffffULL))), "87ffffffffffffff");
 }
 
 TEST(RLP, uint256_t) {
